only killall server if running

+++ scripts/stop_server.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 
-killall server
+if pidof server; then
+    killall server
+fi
